refactor: 資格試験更新部分を切り出し
