+++ workspaces/client/app/components/feature/user/setting/hooks/use-delete-user-certification.ts
@@ -1,14 +1,6 @@
-import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
+import { useMutation, useQueryClient } from "@tanstack/react-query";
 import { useRepository } from "~/hooks/use-repository";
 import { useToast } from "~/hooks/use-toast";
-
-export const useCertifications = () => {
-	const { certificationRepository } = useRepository();
-	return useQuery({
-		queryKey: certificationRepository.getAllCertifications$$key(),
-		queryFn: certificationRepository.getAllCertifications,
-	});
-};
 
 export const useDeleteUserCertification = () => {
 	const queryClient = useQueryClient();
