+++ workspaces/client/app/routes/dashboard/settings/page.tsx
@@ -1,6 +1,6 @@
 import type { MetaFunction } from "react-router";
 import { css } from "styled-system/css";
-import { CertificationRequest } from "./internal/components/certification-request";
+import { CertificationRequestDialog } from "~/components/feature/user/setting/certification-request-dialog";
 import { ProfileUpdateForm } from "./internal/components/form";
 import { ProfileImageEditor } from "./internal/components/profile-image-editor";
 
@@ -44,7 +44,7 @@ export default function Settings() {
 				<ProfileImageEditor />
 				<ProfileUpdateForm />
 			</div>
-			<CertificationRequest.Root />
+			<CertificationRequestDialog.Root />
 		</div>
 	);
 }
